requirements.txt
azure-core==1.39.0
azure-identity==1.25.3
azure-keyvault-secrets==4.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
greenlet==3.3.2
idna==3.11
isodate==0.7.2
msal==1.35.1
msal-extensions==1.3.1
numpy==2.2.6
pandas==2.3.3
pyarrow==23.0.1
pycparser==3.0
PyJWT==2.12.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
